requirements.txt
asgiref==3.3.1
cffi==1.14.5
cryptography==3.4.6
Django==3.1.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.12.2
importlib-metadata==3.4.0
Markdown==3.3.3
psycopg2==2.8.6
pycparser==2.20
pytz==2020.5
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
